Move VQEAdapt to qiskit/chemistry/aqua_extensions
This change is necessary since VQEAdapt depends on UCCSDAdapt which in
turn depends on UCCSD which is part of the aqua_extensions module.
If VQEAdapt were to stay within the qiskit.aqua module a cyclic
dependency between qiskit.aqua and qiskit.chemistry would be introduced
which is definitely NOT desirable.