FCIDump Driver (Qiskit/qiskit-aqua#859)
* Implement FCIDumpDriver

The FCIDumpDriver provides the means to construct a QMolecule instance
out of a FCIDump file.
It is split into three parts:
  - the `driver` itself which provides the normal run() method
  - a separate `parser` implementation which handles FCIDump reading
  - an additional `dumper` implementation which allows FCIDump writing

* Add FCIDumpDriver unittests

* Fix typo and remove from pylintdict

* Fix copyright year

* Fix copyright year - part 2

* Finally, make copyright correct.

The copyright should include:
  <year of file creation>, <year of last modification>
In the case were these are the same, it is sufficient to state it once.

* Refactor FCIDump parsing unittests

Instead of relying on a hard-coded RHF and UHF testcase for H2, we are
now testing against the FCIDump files for H2, LiH and OH.
For the latter two cases the numpy reference arrays are provided as
compressed archive files.

* Store inactive energy in nuclear repulsion energy instead of HF energy.

* Add FCIDumpDriver unittest for stack integration

* Check for num_atoms in QMolecule.core_orbitals()

In the example of a `freeze_core` Hamiltonian this method is called to
return the list of core orbitals.
When working with the FCIDumpDriver it may be the case that `num_atoms`
is not set. Thus, we check for this case and return in empty list
instead.

Unittests have been added which assert that a warning is logged.

* Add optional num_particles argument

Allow the user to manually specify the number of particles.

* Add optional atoms argument

Allow the user to manually specify the molecule information.
This permits frozen core Hamiltonians to be created from
FCIDumpDriver-generated QMolecule instances.

* Update FCIDumpDriver unittest documentation

* Add MS2, ISYM and ORBSYM optional arguments

* Add unittest for dumping RHF H2 QMolecule instance

* Use MS2 (spin quantum number) properly

Instead of the optional `num_particles` spin quantum number from the
FCIDump (MS2) can be used. This gives the multiplicity directly.

* Update copyright

* Ignore possible MO energy values

This will prevent the parsing from failing if the MO energies are in
fact present.

* Remove too specific pylintdict exceptions

* Improve error messages

* Minor code improvements

* Fix docstrings and type hints.

* Ensure QMolecule.log() works with FCIDumpDriver

Also adds unittests.

* Use typing types instead of builtins

* Fix typo

* Remove Optional type hint where not applicable

* Raise error if required field is missing in FCIDump

* Revert to None as default and check in QMolecule.log()