Add basic ParticleOperator interfaces
This poses the start of a major refactoring of Qiskit Chemistry's input
flow. Our goal is to streamline everything from the output of the driver
to the construction of the qubit operator which will be used to solve a
chemistry-related problem.

We have to the conclusion, that the most natural representation and the
smallest common denominator of chemistry (and possibly physics) related
problems are Hamiltonian in the second quantized formalism. Thus, we are
introduced the `SecondQuantizedOperator` as a base-class for all problem
specifications prior to being mapped to a qubit-level operator.

On top of this `SecondQuantizedOperator` we will implement general
operator transformations such as reductions to active spaces or
seniority zero treatments (which change the size of the Hilbert space of
the problem) as well as transformations such as the
Particle/Hole-transformation which only changes the physical context in
which the problem is viewed.
Furthermore, we will re-implement the Fermion-to-Qubit mappings as
standalone classes in order to be more in-line with the rest of Qiskit's
object-oriented implementation as well as ease future extensions.

To this extent, in this commit we introduce the basic
`SecondQuantizedOperator` and some further operator interfaces which
will be combined in this operator. The `FermionicOperator` and
`BosonicOperator` represent basic creation- and annihilation-operators
from the second quantization which act on fermions or bosons,
respectively.
In addition, the `SpinOperator` provides an implementation of
arbitrary-spin particles based on the mathematical foundation as
outlined (e.g.) in [1]. These will likely mainly be used in
physics-related problems but will be integrated into the
`SecondQuantizedOperator` seamlessly.

This work is largely based on the work of Simon Mathis and Giulia
Mazzola.

[1]: https://en.wikipedia.org/wiki/Spin_(physics)#Higher_spins

Co-authored-by: Giulia Mazzola <giuli.mazzola@gmail.com>
Co-authored-by: Simon Mathis <simon.mathis@gmail.com>