Add QubitMapping interfaces and class templates
The `QubitMapping` interface defines the base for how any method which
maps a problem-specific Hamiltonian into a qubit-space Hamiltonian
should be implemented. It provides (for now) one main function called
`map` which should implement this logic and translate a
`SecondQuantizedOperator` into a `PauliSumOp`.

The auxiliary function `supports_particle_type` can be used to check
whether a given particle-type is supported by the mapping.