Add SecondQuantizedTransformation interfaces and class templates
The `SecondQuantizedTransformation` interface defines the base for how
any method which maps one `SecondQuantizedOperator` to another one
should be implemented. This includes such transformations which (as a
side effect or maybe even sole goal) reduce the size of the underlying
Hilbert space (such as e.g. the active-space reduction).
For now, the interface provides one main function called `transform`
which should implement the logic of the transformation. If some setup is
required by the user this should be done in the algorithms initiation as
show-cased in the `ActiveSpaceTransformation` class.

At this time it is not clear whether we need to support a common
side-effect mechanism. This would be useful to centrally handle side
effects such as energy offsets which need to be accounted for in the
problem's total energy. Time will tell whether we can find a common
solution within the interface or whether this needs to be done by the
user manually (unless used through a wrapper such as the (planned)
`ProblemTranslation` interface, which will supersede the
`FermionicTransformation`).