Improve JW mapping logic
We must respect right-to-left order execution of operators.
When doing so, we (obviously) must not reverse the order in which the
fermionic label is iterated such that we end up with the more natural
`NIII`-like labels rather than `EIII`-kind.
This is important as it is the more natural way to think about these
fermionic labels.