Make energy_shift a dictionary
In order to be able to preserve the energy shifts introduced by
different transformers, we can store them in a dictionary until we
refactor the driver and QMolecule classes. It is then easy to sum over
the `dict.values()` to obtain the total energy shift while also
retaining the information which transformer introduced what amount.