Implement the ActiveSpaceTransformer (#50)
* Adapt new transformers to new design

* Port active-space transformer from dev branch

Note: the ActiveSpaceTransformer was developed internally over an
extended period of time. However, due to the move to a new repository
and the extensive code refactoring, I cannot preserve the development
history of this class. Thus, it is now added in bulk.

* Fix copyright years in newly created files

* Do not use bare assert statements

* Add required occupation number information to QMolecule

* Add some unittests for the ActiveSpaceTransformer

* Fix mypy

* Fix spell

* Extract complex logic into auxiliary methods

* Make energy_shift a dictionary

In order to be able to preserve the energy shifts introduced by
different transformers, we can store them in a dictionary until we
refactor the driver and QMolecule classes. It is then easy to sum over
the `dict.values()` to obtain the total energy shift while also
retaining the information which transformer introduced what amount.

* Also validate the number of active electrons in custom AS choices

* Improve docstring

* Fix mypy

* Fix spell

* Handle dict type during QMolecule.save

* Parametrize unittest

* Load test data from constant hdf5 file

* Handle missing `mo_occ` case

Since we will refactor the drivers at a later point in time, we must
support a scenario in which `mo_occ` is not set in the `QMolecule`
class. In this case, we construct the occupation vector based on the
assumption of a HF ground state.

* Use transformer class name as energy_shift key

* Add missing HDF5 test file