Remove AdaptiveAnsatz interface
Rather than duplicating the adaptive nature of an Ansatz through the
introduction an additional interface we can rely on an adaptive
algorithm to handle the list of excitations. Currently (AdaptVQE) does
this anyways and the `push` and `pop` methods were merely utilities.
Now, the BlueprintCircuit can simply have its operators replaced which
will automatically invalidate a cached circuit and will cause rebuilding
of all relevant information.