Remove SecondQuantizedOp class
Based on an offline discussion, we have decided to remove the
SecondQuantizedOp class as it stands now. For current applications it is
mainly a container which everything needs to be placed without adding
useful features.
Once we need a mixed-particle kind of operator, we can implement a
`MixedOperator` which will be a "composite"-like implementation of the
current `ParticleOp` interface. This will allow seemless usage of the
mixed- as well as pure-particle operators within the stack.

However, the addition of this class is postponed until further notice.