Extract commutation lookup
Rather than duplicating the method calls with different arguments, the
logic of determining the correct keyword arguments for the update
methods has been extracted.
This is a significant and very useful refactoring except for one
remaining issue: the hard-coded list of anti-commutation rules in the h2
case. For now, there is no time to properly implement this in an
operator-based fashion, but this may be an improvement we can look into
in the future.
Nonetheless, we should investigate the possibility of a more suitable
data structure to perform the lookup (see also the inlined TODO).