Replace `pure_spin` with `max_spin_excitation`
The original intent of `pure_spin` was to add an easy configuration
option to specify the SUCC Ansatz. However, the original implementation
was backwards: `pure_spin` removed the alpha,beta double excitations but
in fact, we want to remove the alpha,alpha and beta,beta ones.

In the new implementation we have the optional integer argument called
`max_spin_excitation`. This can be used to set the
    maximum number of excitations within identical spin orbitals!

For example, setting `num_excitations=2` will compute all possible
double excitations. However, in combination with `max_spin_excitation=1`
this list will be filtered as to only include those where per spin, we
perform at most 1 excitation.