ActiveSpaceTransformer generalization (#77)
* Remove FreezeCoreTransformer template

* Remove left-over test init file

* WIP: extend ActiveSpace to handle dipole operators

* Generalize ActiveSpaceTransformer for dipoles

The dipole integrals must also be reduced into the selected active space
in accordance with the electronic energy operator. This can easily by
done by reusing the functionality without using the electron repulsion
integrals.

* Use subTest and test dipole integrals

* Also load the energy shift dicts from an HDF5 file

* Add `freeze_core` convenience argument

It can be convenient to simply remove the core orbitals supplied by the
`QMolecule` class. To achieve this, the `freeze_core` option can be
enabled. This will overwrite all other arguments.

* Test freeze_core option

* Remove code duplication

* Extract large integral matrices into HDF5 file

* Fix spell

* Fix mypy

* Fix MolecularProblem unittests

Reasons for failing:
- the number of expected operators increased since the
  ActiveSpaceTransformer now provides dipole integrals
- the expected integral values were updated

* Disambiguate `num_orbitals`

* Document requirement of AO-basis matrices

* Fix lint

* Fix spell

* Add `remove_orbitals` kwarg to ActiveSpaceTransformer

Since it is so straight forward to add the ability to remove virtual
orbitals in this new implementation, I decided to add another keyword
argument which can be used in combination with `freeze_core`.

* Fix style

* Rename internal variable num_orbitals -> num_molecular_orbitals

* Extract FreezeCoreTransformer

* Refactor arguments into single `num_electrons`

* Update docstring

* Fix test import

* Fix style