Actually leverage QubitConverter during algorithm
It does not make sense to not use the conversion logic of the
QubitConverter during the algorithm. We can still do an additional
`mapping` step (although this is not very performant) but if we were to
do every symmetry reduction (currently missing the 2-qubit reduction) on
its own, we don't gain anything from having the converter object.