Remove ooVQE for upcoming release
ooVQE is plagued by some problems which we cannot resolve cleanly before
the upcoming release. Thus, we will remove it entirely and introduce it
in the future, after it has been refactored and improved considerably.

The most optimal version of the implementation can be found in
23e1fc77347968c1ae7bdfd3f5ce264ca705bf6a.
That commit also introduced a TODO list with some issues which I will
reproduce here for convenience:

- removal of `molecular_problem` from `__init__`
  (this will only be given to `solver()`
- actual use of `MolecularProblem.second_q_ops()` rather than
  hard-coded `build_fermionic_op`
- unification with `BaseProblem.solve()`
- general refactorings of flow
- fix freeze-core support
- fix dependence on `num_parameters` handling within Terra's VQE
  and Optimizer