Revert "Extend PySCFDriver to allow DFT calculations"
This reverts commit 3611f399f95145a960c9ede65fbf9d9eadb7f91a.

The reason this is necessary is that the DFT-case requires a more care.
What the driver prior to this reversion would have done is a HF-style
calculation using DFT-orbitals. But the XCF would not be contained in
the Hamiltonian because the driver simply uses `get_hcore()` (just as in
the HF-case). Instead, only `get_fock()` actually contains the
DFT-XC-parts but it cannot simply replace `get_hcore()` because it also
contains the Coulomb and HF-Exchange terms.
I propose to finish this HF-case driver as is and add a derived
`PySCFDFTDriver` later, instead of trying to integrate them.