Re-add deprecated code
As pointed out by Steve it makes more sense to deprecate the old code in-place since we are planning
more refactorings for the code which got moved. Deprecating the old code in-place ensures a proper
deprecation without relying on the upcoming refactorings to be backwards compatible (which they are
most likely not).