PySCF Driver Updates (#180)
* Extend PySCFDriver to allow DFT calculations

* Integrate pyscfd/integrals.py into driver file

The motivation for doing this is to allow an even more flexibel
PySCFDriver implementation in which an advanved user has the ability to
monkeypatch the method which performs the actual calculation of a
specific instance at runtime.

* Expose chkfile option in PySCFDriver

* Leverage class variables

* WIP: extract actual PySCF calculation into separate function

* Store mol and m_f in object

* Revert "Extend PySCFDriver to allow DFT calculations"

This reverts commit 3611f399f95145a960c9ede65fbf9d9eadb7f91a.

The reason this is necessary is that the DFT-case requires a more care.
What the driver prior to this reversion would have done is a HF-style
calculation using DFT-orbitals. But the XCF would not be contained in
the Hamiltonian because the driver simply uses `get_hcore()` (just as in
the HF-case). Instead, only `get_fock()` actually contains the
DFT-XC-parts but it cannot simply replace `get_hcore()` because it also
contains the Coulomb and HF-Exchange terms.
I propose to finish this HF-case driver as is and add a derived
`PySCFDFTDriver` later, instead of trying to integrate them.

* Auto-format PySCFDriver

* WIP: PySCFDriver initializer updates + properties

* Split into utility methods

* Update docstrings

* Add DFT methods for KS-orbital based calculations

After a back-and-forth on the addition of the DFT methods, this commit
finally re-adds them as an option. However, the user should be warned as
this option merely serves the purpose of using KS-orbitals for an
otherwise unchanged post-HF calculation of Qiskit.

* Fix error raising docstrings

* Fix spell

* Catch invalid input type

* Fix lint

* Update docstrings

* Fix mypy for Python 3.7

On Python 3.9 mypy already passed because the assignment type is already
being ignored on the correct line number. However, Python 3.7 apparently
does not detect that correctly and requires this additional
`type: ignore` comment.

* Basic PySCF Driver updates

- the "methods" naming is non-descriptive and misleading. It is changed to "problem" to reflect the
  fact that these tests are to ensure the drivers work correctly within the Nature stack around the
  Problem and Algorithm instances.
- the "extra" tests are moved to the standard class for simplicity. The OH cases are marked as slow.

Disclaimer: non-PySCF related updates will be tackled later

* Fix test suite after rebase

* Fix driver changes after rebase

* Remove slow_test label from OH tests

* Address more review comments

* Fix lint

* Fix PySCF installation link

* Revert unittest naming changes

* Fix lint