Revert ListOrDict integration
We need to properly deprecate the old function signature of
`second_q_ops` before we can fully switch to dict-based aux operators.
For this, I introduce a new keyword argument `return_list` which
defaults to the old way of list-based aux operators.

In the following commits I will add DeprecationWarnings announcing a
change of this default as well as unittests to assert the correct
behavior of the dict-based aux operators.