Rename to more general interaction_matrix
Lattice models in general do not have a "hopping" matrix. This is a
rather specific interpretation which makes sense for the
`FermiHubbardModel`. In the case of the `IsingModel`, it can be
interpreted as a coupling matrix.
Thus, the variable name is changed to `interaction_matrix` which is
model agnostic for the interface. For convenience, the various models
can then expose this matrix via reasonably named property attributes.