Update `load_from_hdf5` (#554)
* refactor: return first object in load_from_hdf5

The documentation of the `load_from_hdf5` suggested that one could
immediately obtain a driver result from it. However, in reality, this
would have assigned a generator method instead of the actual result.
In order to be aligned with the `save_to_hdf5` implementation, this
commit changes the `load_from_hdf5` function to return the first
constructed object and ignore any further ones. In practice this should
be sufficient because the Qiskit Nature stack is written such that only
a single object will be stored per HDF5 file.

* test: add HDF5 method unittests

* fix: Windows CI file permissions

* test: add load_from_hdf5 exception tests