fix: catch DriverMetadata-related warnings during transform
The DriverMetadata property currently raises a lot of errors when being
transformed. Since this part of the normal stack this should be avoided.
For the time being, we simply catch the case and pass the DriverMetadata
object down to the transformed grouped property unchanged. However, in
the future we could envision storing additional information about the
transformer in the DriverMetadata container.