Allow driver result to be passed to solver
The changes in this commit allow for a previously calculated driver
result to be passed to the eigensolver for it to get information on the
molecule. In its current state, the driver is run regardless of whether
or not the result is already known.

As the size of the models increase, it is important that expensive operations
like driver.run() are used only when necessary and avoided when
possible. This change introduces the framework for that.