refactor: update HDF5 files
This updates the contents of the HDF5 files used during unittesting.
Since this is a temporary migration of all the existing properties into
a new module location, this means that the module import paths stored
inside of the HDF5 files are being updated by this PR. Nothing else
changes functionally.