refactor: Towards a non-driver future
As outlined in #701, Qiskit Nature is working towards a non-driver
focused future. This means, rather than the `BaseProblem` interface
taking in a `BaseDriver` and an optional list of `BaseTransformer`
instances, the problem itself should be a fully standalone description
of the user's problem of interest.
In other words, `BaseProblem` should be the produced result of a
`BaseDriver` and, by extension, `BaseTransformer` instances should
transform problem instances.

This commit works towards such a future by refactoring the code:
- `BaseProblem` is the returned object of `BaseDriver.run`
- `BaseTransformer.transform` has `BaseProblem` as in- and output
- `BaseProblem` takes over the role of `GroupedProperty`
  - in doing so, `Hamiltonian` objects are more clearly separated from
    `Property` objects (further refactoring to come)