feat: support basic QCSchema
Adds dataclasses to represent data stored in the QCSchema format [1].
Some basic test cases have been taken from the QCSchema test files.

For this to be usable, further changes are required extending the
QCSchema to support active-spaces, by including the two-body terms
directly in the files.
Support for HDF5 files will also need to be added.

[1]: https://molssi-qc-schema.readthedocs.io/en/latest/index.html