feat: PropertiesContainer class
This adds the PropertiesContainer class which is a storage object for
Property instances. The key feature is that it only allows a single
instance of any Property kind to be stored at any time.
This is sufficient for all application purposes since all auxiliary
operators generated by the Property instances must match in dimension
anyways.