refactor: move QEOM hopping op builders
This moves the QEOM hopping operator building utilities away from the
`BaseProblem` interface and integrates them into the `QEOM` algorithm
itself. This is a step towards the leaner `BaseProblem` interface which
we are targeting during the ongoing refactoring.
This also decouples the problem classes from a very specific algorithm.
Instead, the algorithm itself is now in charge of adding support for
other problem types.