refactor: use TemporaryDirectory where possible
Instead of using a NamedTemporaryFile and juggle its existence state,
this commit refactors all scenarios to use a TemporaryDirectory within
which the file names can be controlled more easily.
The reason for this refactoring was to guard against possible runtime
conflicts when multiple processes might be attempting to access the same
temporary file name, after the object has been unlinked (which might
free up the temporary name for reuse).