PropertiesContainer (#766)
* feat: PropertiesContainer class

This adds the PropertiesContainer class which is a storage object for
Property instances. The key feature is that it only allows a single
instance of any Property kind to be stored at any time.
This is sufficient for all application purposes since all auxiliary
operators generated by the Property instances must match in dimension
anyways.

* fix pylint

* fix docs

* fix: type in releasenote

* Add Terra intersphinx and restore the sys.path in conf.py as before.

* refactor: extract attribute setter implementation

This extracts the identical logic for all attribute setters into a
single, central location.

* docs: update PropertiesContainer attribute getter docstrings

* refactor: extract attribute getter implementation

Co-authored-by: Manoel Marques <manoel.marques@ibm.com>
Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>