WIP: problem refactoring
This is a first step towards decoupling the BaseProblem and BaseDriver
interfaces. Instead of injecting drivers and transformers into a
problem, these become the output of our driver classes. In turn,
transformers now take entire problem instances as input and return them
as output.

This is an intermediate step towards the final form of the problem
classes and this commit is not working fully.