Fix #788
This commit properly ensures the compatibility of the generalized and
preserve_spin keyword arguments available in the fermionic excitation
generator methods.
A previous attempt to fix this was done in #747, but some edge cases
were overlooked.

This commit also includes the same patch as a bugfix in the legacy code
location.