refactor: handle subclass keyword arguments of SparseLabelOp
This enables the handling of non-default keyword arguments in subclasses
of the SparseLabelOp (for example the spin in SpinOp and num_modes and
num_modals in VibrationalOp). To do so, it even removes the
register_length argument from the base class in favor of the dedicated
num_spin_orbitals argument on the FermionicOp subclass.