feat: custom dipole support in QCSchema
This adds custom support to store dipole moment integrals in the
QCSchema format. While this is not part of the official specification,
it is aligned well with the remainder of the QCSchema and is not the
only custom extension which we support.
This will allow us to achieve a more consistent and simple
implementation of our classical drivers.