docs: document inherited members in mappers module
This is more of an ad-hoc fix for the current situation. Qiskit Terra
also already differentiates in cases where they want to document the
inherited members and where they do not. However, they default to
document them and disable them as needed. For the scope of this PR, I
chose to do the inverse but we should do a follow-up and align our
documentation style with that of Qiskit Terra in the future.