Some minor typing fixes discovered via pyright
In general pyright seems to find quite a few "errors" which mypy seems
to overlook. On the other hand pyright also seems to raise a lot of
false positives in this repo some of which might be related to mixins
coming from Terra, others related to dunder methods or untyped
dependencies. I did not do a deep dive into all of the raised issues so
this commit simply fixes a few "major" ones.