docs: fix flakiness of AdaptVQE guide
doctest appears to execute all gathered code blocks in one Python
session. Since the order of the executed files is not guaranteed to be
the same, this can cause flaky failures when one file changes settings
which affect another one.
This commit prevents that by enforcing a given setting as long as Qiskit
Terra 0.24.0 is not released yet.