feat: improve the implementation of the AngularMomentum operator
The AngularMomentum operator is actually just the $S^2$ operator.
However, in the code prior to this commit this was really not easy to
follow.

Instead, this commit changes the implementation to rely on the
sub-operators which the $S^2$ operator consists of. More specifically,
it adds generator functions for the $S^+$ and $S^-$ ladder operators as
well as the $S^x$, $S^y$, and $S^z$ operators. All of these may also be
useful to users who want to evaluate these separately.

Unittests are added to assert the relations of all of these operators.