UCC and fully occupied MOs (Fix for issue #1277)
The configuration check looked for whether any number of alpha or beta orbitals was greater or equals than the number of particles.
Obviously, there cannot be more particles than spin orbitals, and in order to handle this case, I updated the previous `any(n >= ...` to `any(n > ...`. 
For handling case where both alpha and beta are the same as the number of particles (i.e. He2 molecule), I created a new raise condition, notifying that the function is still not implemented.
For the case where either alpha or beta is equals to the number of particles (i.e. H-He molecule), it should just pass through.