ci: update deploy-docs authorized users (#1283)
In order to deploy the documentation to the public URL, a Github action
needs to be triggered manually. In order to avoid anyone being able to
do this, the script which is being run asserts the username of the
`github.actor` to be in a hard-coded list of authorized users.

This commit updates this list in accordance with the recent CODEOWNER
changes.

I believe, that there must be a better way to handle/restrict access to
who can trigger this action, but for the time being I will simply do
this quick fix.

(cherry picked from commit 4ce231e7acd00136b75bd4ecaedd9c45e3e9af89)