UCC and fully occupied MOs (Fix for issue #1277) (#1286)
* UCC and fully occupied MOs (Fix for issue #1277)

The configuration check looked for whether any number of alpha or beta orbitals was greater or equals than the number of particles.
Obviously, there cannot be more particles than spin orbitals, and in order to handle this case, I updated the previous `any(n >= ...` to `any(n > ...`. 
For handling case where both alpha and beta are the same as the number of particles (i.e. He2 molecule), I created a new raise condition, notifying that the function is still not implemented.
For the case where either alpha or beta is equals to the number of particles (i.e. H-He molecule), it should just pass through.

* Remove unintended spaces

Sorry, I forgot to remove the spaces.

* Added case for HeH 

Added a test case for HeH molecule, with (alpha, beta) = (2, 1). This should work after the changes in ucc.py

* Remove comment

The tests complained because of the spelling of a molecule in a comment. And I reformatted a little.

* Formatting according to black

Formatted the indentations correctly following the suggestions of black

* Create release notes for PR #1286

Release notes created

* Deleted unnecessary parts.

Deleted unnecessary parts.

* Update releasenotes/notes/fix-UCC-and-fully-occupied-MOs-c8fdf252a0607395.yaml

---------

Co-authored-by: Max Rossmannek <max.rossmannek@uzh.ch>